refactor: activity, application module을 분리한다.
