+++ di/src/main/java/com/boogiwoogi/di/Module.kt
@@ -5,7 +5,7 @@ import kotlin.reflect.KClass
 /**
  * the class that provide how to instantiate some classes
  */
-interface Modules {
+interface Module {
 
     fun provideInstanceOf(clazz: KClass<*>): Any?
