--- woogi-di/src/main/java/com/boogiwoogi/di/DiComponent.kt
@@ -1,6 +1,0 @@
-package com.boogiwoogi.di
-
-interface DiComponent {
-
-    val instanceContainer: InstanceContainer
-}
